Ledge Detection
Created a copy of the template scene in my personal folder (Vihaan). Used transform.Translate to move the enemy so as to not mess with the physics system by using a rigidbody. When the enemy detects a ledge, it simply turns and moves in the opposite direction until it finds another ledge. Also, added methods to show where the enemy is grounded at any given time and also gives its exact position at any given time.