NavMesh2DImplementation
Implemented NavMesh 2D using Unity.AI which allows all enemy players to avoid the obstacles that are placed in the scene. NavMesh is baked at Compile time and can adapt to the scene as well.

I add two new scripts, NavMeshAgent (which represents the enemy) and RenderingOffset which solves the Offset problem Alex and I were facing.